+++ home-manager/default.nix
@@ -2,6 +2,7 @@
   imports = [
     ./modules/git.nix
     ./modules/gpg.nix
+    ./modules/identity.nix
     ./modules/zsh.nix
   ];
 }